Manual

FeatureEng.jl

Installation

Quick Start

Manual

Categorical Features

DateTime Features

Numeric – Binning Features

Numeric – Scaling Features

Numeric – Transforming Features

Numeric – Interaction Features

Categorical Features

FeatureEng.encode_dummy — Method

encode_dummy(column::T[, categories::T[, prefix::String]]) where T <: AbstractArray

Same as encode_onehot except that it drops the first column (to help prevent issues caused by multicollinearity).

See also: encode_onehot, encode_hash

FeatureEng.encode_hash — Method

encode_hash(column::T, n_cols::Int = 8, prefix::String = "c") where T <: AbstractArray

Deterministically encode categorical features with high cardinality as a DataFrame with n_cols columns.

data = [1:100;1000;];

encode_hash([1:1_000:10_000;])

See also: encode_onehot, encode_dummy

FeatureEng.encode_onehot — Method

encode_onehot(column::T[, categories::T[, prefix::String]]) where T <: AbstractArray

Converts a categorical column into a DataFrame of one-hot-encoded columns – with one binary-encoded column per unique value in column.

Examples

The basic version of this function makes a column for each unique value in column.

data = [3,1,2,4];

encode_onehot(data)

You can also specify a prefix for each column.

data = [3,1,2,4];

encode_onehot(data,prefix="col_")

Additionally, you can specify the categories to convert to columns, regardless of whether it exists in column.

data = [3,1,2,4];

encode_onehot(data,[1:6;],"c")

See also: encode_dummy, encode_hash

DateTime Features

FeatureEng.extract_date_features — Method

extract_date_features(datetimes::T) where T <: AbstractArray{<:Union{Date,DateTime}}

Extract a DataFrame of features from an array of DateTime or Date objects. Features extracted:

year: Year from datetime

month: Month from datetime

dayofmonth: Day of the month (1-31)

dayofweek: Day of the week (ordered)

isweekend: Is datetime a weekend?

quarter: The quarter from datetimes

Examples

data = strp_datetimes([
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "2010-06-11 11:00:00"
]);
extract_date_features(data)

See also: extract_datetime_features, extract_time_features

FeatureEng.extract_datetime_features — Method

extract_datetime_features(datetimes::T) where T <: AbstractArray{<:DateTime}

Extract a DataFrame of features from an array of DateTime objects. Features extracted:

year: Year from datetime

month: Month from datetime

dayofmonth: Day of the month (0-31)

dayofweek: Day of the week (ordered)

isweekend: Is datetime a weekend?

quarter: The quarter from datetimes

hour: Hour of the day from datetime

minute: Minute from datetime

second: Second from datetime

isAM: Is time AM (vs PM)?

The same as the following:

hcat(
    extract_date_features(datetimes),
    extract_time_features(datetimes)
    )

Examples

data = strp_datetimes([
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "2010-06-11 11:00:00"
]);
extract_datetime_features(data)

See also: extract_date_features, extract_time_features

FeatureEng.extract_time_features — Method

extract_time_features(datetimes::T) where T <: AbstractArray{<:Union{Time,DateTime}}

Extract a DataFrame of features from an array of DateTime or Time objects. Features extracted:

hour: Hour of the day from datetime

minute: Minute from datetime

second: Second from datetime

isAM: Is time AM (vs PM)?

Examples

data = strp_datetimes([
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "2010-06-11 11:00:00"
]);
extract_time_features(data)

See also: extract_datetime_features, extract_date_features

FeatureEng.get_month — Method

get_month(datetimes::T) where T <: AbstractArray{<:Union{Date,DateTime}}

Return an ordered CategoricalArray of month names extracted from datetimes.

Examples:

data = strp_datetimes([
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "2010-06-11 11:00:00"
]);
get_month(data)

See also: extract_datetime_features, extract_date_features, get_weekday

FeatureEng.get_weekday — Method

get_weekday(datetimes::T) where T <: AbstractArray{<:Union{Date,DateTime}}

Return an ordered CategoricalArray of weekday names extracted from datetimes.

Examples:

data = strp_datetimes([
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "2010-06-11 11:00:00"
]);
get_weekday(data)

See also: extract_datetime_features, extract_date_features, get_weekday

FeatureEng.strp_datetimes — Method

strp_datetimes(datetimes::T, format::Union{String,DateFormat} = "y-m-d H:M:S") where T <: AbstractArray{<:AbstractString}

Convert an array of timestamps and to an array of DateTime objects.

Any of the strings it's unable to parse, will be replaced with missing.

Examples

date_strings = [
    "2021-01-27 14:03:25",
    "1999-10-05 01:13:43",
    "abcdefg"
    ];
strp_datetimes(date_strings)

Numeric – Binning Features

Numeric – Scaling Features

Numeric – Transforming Features

FeatureEng.transformBoxCox — Method

transformBoxCox(data::T, λ::Real = 0.0) where T <: AbstractArray{<: Number}

Box-Cox power transformation following the following function:

\[y_i^{(\lambda)} = \left\{\begin{matrix} \frac{y_i^\lambda - 1}{\lambda} & \mathrm{if} \lambda \neq 0, \\ \mathrm{ln} y_i & \mathrm{if} \lambda = 0, \end{matrix}\right.\]

Examples:

data = [0:5;];
transformBoxCox(data)
transformBoxCox(data,.1)
transformBoxCox(data,1)

See also: transformLog, transformRoot

FeatureEng.transformLog — Method

transformLog(data::T, base::Real = ℯ) where T <: AbstractArray{<: Number}

Log transform data using log-base, base.

Examples:

data = [0:5;];
transformLog(data)
transformLog(data,2)

See also: transformRoot, transformBoxCox

FeatureEng.transformRoot — Method

transformRoot(data::T, index::Real = 10) where T <: AbstractArray{<: Number}

Root transform data using root index, index.

Examples:

data = [0:5;];
transformRoot(data)
transformRoot(data,2)

See also: transformLog, transformBoxCox

Numeric – Interaction Features

FeatureEng.polynomial — Method

polynomial(df::DataFrame, degree::T = 2) where T <: Integer

Calculate polynomial interaction terms between columns in a DataFrame.

If you have a DataFrame with 3 columns: x, y, and z, you can get degree-2 polynomial interaction terms: x*x, x*y, x*z, y*y, y*z, and z*z.

Examples

using DataFrames
df = DataFrame(a=1:10,b=repeat(0:1,5))
polynomial(df,2)
polynomial(df,3)